Use internal when possible, move a few api/components classes into api/components/data
