+++ src/main/java/com/freya02/botcommands/api/pagination/interactive/InteractiveMenu.java
@@ -1,6 +1,6 @@
 package com.freya02.botcommands.api.pagination.interactive;
 
-import com.freya02.botcommands.api.components.InteractionConstraints;
+import com.freya02.botcommands.api.components.data.InteractionConstraints;
 import com.freya02.botcommands.api.pagination.TimeoutInfo;
 import com.freya02.botcommands.api.utils.ButtonContent;
 import net.dv8tion.jda.api.interactions.components.selections.StringSelectMenu;
