+++ src/main/java/com/freya02/botcommands/api/pagination/menu/ChoiceMenu.java
@@ -1,6 +1,6 @@
 package com.freya02.botcommands.api.pagination.menu;
 
-import com.freya02.botcommands.api.components.InteractionConstraints;
+import com.freya02.botcommands.api.components.data.InteractionConstraints;
 import com.freya02.botcommands.api.pagination.ButtonContentSupplier;
 import com.freya02.botcommands.api.pagination.PaginatorSupplier;
 import com.freya02.botcommands.api.pagination.TimeoutInfo;
