+++ src/main/kotlin/com/freya02/botcommands/api/components/builder/ITimeoutableComponent.kt
@@ -1,6 +1,6 @@
 package com.freya02.botcommands.api.components.builder
 
-import com.freya02.botcommands.internal.components.data.ComponentTimeout
+import com.freya02.botcommands.api.components.data.ComponentTimeout
 import kotlinx.coroutines.runBlocking
 import java.util.concurrent.TimeUnit
 import kotlin.time.Duration
