+++ src/main/kotlin/com/freya02/botcommands/api/components/data/ComponentTimeout.kt
@@ -1,4 +1,4 @@
-package com.freya02.botcommands.internal.components.data
+package com.freya02.botcommands.api.components.data
 
 import kotlinx.datetime.Instant
