+++ src/main/kotlin/com/freya02/botcommands/api/components/data/ComponentTimeoutData.kt
@@ -1,3 +1,3 @@
-package com.freya02.botcommands.api.components
+package com.freya02.botcommands.api.components.data
 
 data class ComponentTimeoutData(val componentId: String)
