+++ src/main/kotlin/com/freya02/botcommands/internal/components/controller/ComponentTimeoutManager.kt
@@ -1,11 +1,11 @@
 package com.freya02.botcommands.internal.components.controller
 
-import com.freya02.botcommands.api.components.ComponentTimeoutData
-import com.freya02.botcommands.api.components.GroupTimeoutData
+import com.freya02.botcommands.api.components.data.ComponentTimeout
+import com.freya02.botcommands.api.components.data.ComponentTimeoutData
+import com.freya02.botcommands.api.components.data.GroupTimeoutData
 import com.freya02.botcommands.api.core.annotations.ConditionalService
 import com.freya02.botcommands.api.core.config.BCoroutineScopesConfig
 import com.freya02.botcommands.internal.components.ComponentType
-import com.freya02.botcommands.internal.components.data.ComponentTimeout
 import com.freya02.botcommands.internal.components.data.EphemeralTimeout
 import com.freya02.botcommands.internal.components.data.PersistentTimeout
 import com.freya02.botcommands.internal.components.new.ComponentGroupData
