+++ src/main/kotlin/com/freya02/botcommands/internal/components/data/ComponentData.kt
@@ -1,10 +1,10 @@
 package com.freya02.botcommands.internal.components.new
 
-import com.freya02.botcommands.api.components.InteractionConstraints
+import com.freya02.botcommands.api.components.data.ComponentTimeout
+import com.freya02.botcommands.api.components.data.InteractionConstraints
 import com.freya02.botcommands.internal.components.ComponentHandler
 import com.freya02.botcommands.internal.components.ComponentType
 import com.freya02.botcommands.internal.components.LifetimeType
-import com.freya02.botcommands.internal.components.data.ComponentTimeout
 
 internal abstract class ComponentData(
     val componentId: Int,
