+++ src/main/kotlin/com/freya02/botcommands/internal/components/data/ComponentGroupData.kt
@@ -1,8 +1,8 @@
 package com.freya02.botcommands.internal.components.new
 
+import com.freya02.botcommands.api.components.data.ComponentTimeout
 import com.freya02.botcommands.internal.components.ComponentType
 import com.freya02.botcommands.internal.components.LifetimeType
-import com.freya02.botcommands.internal.components.data.ComponentTimeout
 
 internal class ComponentGroupData internal constructor(
     groupId: Int,
