+++ src/main/kotlin/com/freya02/botcommands/internal/components/data/EphemeralTimeout.kt
@@ -1,8 +1,9 @@
 package com.freya02.botcommands.internal.components.data
 
+import com.freya02.botcommands.api.components.data.ComponentTimeout
 import kotlinx.datetime.Instant
 
-class EphemeralTimeout(
+internal class EphemeralTimeout(
     override val expirationTimestamp: Instant,
     val handler: (suspend () -> Unit)?
 ) : ComponentTimeout
