+++ src/main/kotlin/com/freya02/botcommands/internal/components/repositories/ComponentRepository.kt
@@ -1,14 +1,14 @@
 package com.freya02.botcommands.internal.components.repositories
 
 import com.freya02.botcommands.api.components.Components
-import com.freya02.botcommands.api.components.InteractionConstraints
 import com.freya02.botcommands.api.components.builder.BaseComponentBuilder
 import com.freya02.botcommands.api.components.builder.ITimeoutableComponent
 import com.freya02.botcommands.api.components.builder.group.ComponentGroupBuilder
+import com.freya02.botcommands.api.components.data.ComponentTimeout
+import com.freya02.botcommands.api.components.data.InteractionConstraints
 import com.freya02.botcommands.api.core.annotations.ConditionalService
 import com.freya02.botcommands.internal.components.*
 import com.freya02.botcommands.internal.components.controller.ComponentTimeoutManager
-import com.freya02.botcommands.internal.components.data.ComponentTimeout
 import com.freya02.botcommands.internal.components.data.EphemeralTimeout
 import com.freya02.botcommands.internal.components.data.PersistentTimeout
 import com.freya02.botcommands.internal.components.new.*
