+++ src/main/kotlin/com/freya02/botcommands/internal/components/repositories/ComponentTimeoutHandlers.kt
@@ -1,7 +1,7 @@
 package com.freya02.botcommands.internal.components.repositories
 
-import com.freya02.botcommands.api.components.ComponentTimeoutData
 import com.freya02.botcommands.api.components.annotations.ComponentTimeoutHandler
+import com.freya02.botcommands.api.components.data.ComponentTimeoutData
 import com.freya02.botcommands.api.core.annotations.BService
 import com.freya02.botcommands.internal.core.ClassPathContainer
 import com.freya02.botcommands.internal.core.requireNonStatic
