+++ src/main/kotlin/com/freya02/botcommands/internal/components/repositories/HandlerContainer.kt
@@ -2,6 +2,6 @@ package com.freya02.botcommands.internal.components.repositories
 
 import kotlin.reflect.KFunction
 
-interface HandlerContainer {
+internal interface HandlerContainer {
     operator fun get(handlerName: String): KFunction<*>?
 }
