+++ src/test/java/com/freya02/botcommands/test/commands/slash/SlashInteractiveMenu.java
@@ -4,7 +4,7 @@ import com.freya02.botcommands.api.commands.application.ApplicationCommand;
 import com.freya02.botcommands.api.commands.application.slash.GuildSlashEvent;
 import com.freya02.botcommands.api.commands.application.slash.annotations.JDASlashCommand;
 import com.freya02.botcommands.api.components.Components;
-import com.freya02.botcommands.api.components.InteractionConstraints;
+import com.freya02.botcommands.api.components.data.InteractionConstraints;
 import com.freya02.botcommands.api.pagination.interactive.InteractiveMenu;
 import com.freya02.botcommands.api.pagination.interactive.InteractiveMenuBuilder;
 import com.freya02.botcommands.api.pagination.interactive.SelectContent;
